+++ pm/CHANGELOG.md
@@ -13,6 +13,11 @@ description: A detailed, chronological record of all changes made to a project, 
 - Fixed missing send message button
 - Added support for redirecting to URLs containing query parameters
 - Unified contact and character usernames
+- Improved button icons
+
+### Removed
+
+- Removed username from "Discover Characters" items
 
 ## [0.1.0-8] - 2025-05-05